dodanie pliku na crud pokoju
- przeniesienie funkcji
- dodanie szkieletu funkcji updateroom
